update storage version + create storage changelog

+++ packages/suite/src/storage/migrations/index.ts
@@ -98,7 +98,9 @@ export const migrate = async (
 
             return cursor.continue().then(addMetadataKeys);
         });
+    }
 
+    if (oldVersion < 16) {
         // object store for send form
         // @ts-ignore sendForm doesn't exists anymore
         db.deleteObjectStore('sendForm');
